+++ catkin_tools_fetch/common/__init__.py
@@ -1,0 +1,2 @@
+"""Module for fetching dependencies."""
+__all__ = ["tools"]
