+++ catkin_tools_fetch/fetch/__init__.py
@@ -1,0 +1,2 @@
+"""Module for fetching dependencies."""
+__all__ = ["dependency_parser", "downloader"]
