+++ catkin_tools_fetch/fetch/dependency_parser.py
@@ -8,7 +8,7 @@ import logging
 from os import path
 from xml.dom import minidom
 
-from catkin_tools_fetch.fetcher.tools import Tools
+from catkin_tools_fetch.common.tools import Tools
 
 log = logging.getLogger('fetch')
