--- catkin_tools_fetch/fetcher/__init__.py
@@ -1,2 +1,0 @@
-"""Module for fetching dependencies."""
-__all__ = ["dependency_parser", "downloader", "tools"]
